✨ [Feat] sidebar 애니메이션 추가
