+++ src/components/LogoutButton.tsx
@@ -1,15 +1,23 @@
 'use client';
 
 import { signOut } from 'next-auth/react';
+import { motion } from 'framer-motion';
+import { itemType } from './layout/SideBar';
 
-export default function LogoutButton() {
+export default function LogoutButton({
+    itemVariants,
+}: {
+    itemVariants: itemType;
+}) {
     return (
-        <button
-            className='py-[19px] px-[19px] text-left'
+        <motion.button
+            key={8}
             onClick={() => {
                 signOut({ callbackUrl: '/' });
-            }}>
+            }}
+            whileHover={{ scale: 1.1 }}
+            variants={itemVariants}>
             로그아웃
-        </button>
+        </motion.button>
     );
 }
