+++ src/components/layout/Header.tsx
@@ -1,42 +1,15 @@
 'use client';
 
-import Link from 'next/link';
-import Image from 'next/image';
-import { useEffect } from 'react';
-import useMenuStore from '@/zustand/menuStore';
-import { RxHamburgerMenu } from 'react-icons/rx';
-import { usePathname } from 'next/navigation';
 import useScrollPosition from '@/hooks/useScroll';
-import Nav from './Nav';
+import Image from 'next/image';
+import Link from 'next/link';
+import { usePathname } from 'next/navigation';
+import SideBar from './SideBar';
 
 export default function Header() {
     const pathname = usePathname();
     const isMatchMain = pathname === '/';
-
     const { scrollPosition } = useScrollPosition();
-
-    // # 메뉴 상태 전역 관리
-    const { isOpen, setIsOpen } = useMenuStore();
-
-    const handleOnClick = () => {
-        setIsOpen();
-    };
-
-    // 스크롤 비활성화
-    useEffect(() => {
-        // # 메뉴 열릴 때 스크롤 비활성화
-        if (isOpen) {
-            document.body.style.overflow = 'hidden';
-            // # 메뉴 닫힐 때 스크롤 활성화
-        } else {
-            document.body.style.overflow = 'auto';
-        }
-
-        // # 컴포넌트 언마운트 시 스크롤 활성화
-        return () => {
-            document.body.style.overflow = 'auto';
-        };
-    }, [isOpen]);
 
     return (
         <header
@@ -50,11 +23,8 @@ export default function Header() {
                         height={36}
                     />
                 </Link>
-                <button className='w-9 h-9' onClick={handleOnClick}>
-                    <RxHamburgerMenu className='mx-auto w-7 h-7 text-[#4C4646]' />
-                </button>
+                <SideBar />
             </nav>
-            {isOpen && <Nav />}
         </header>
     );
 }
